[#187] ✨ feat: PlanController 권한 적용, 여행 코스 등록 유효성 검증

+++ src/main/java/team/seventhmile/tripforp/domain/plan/service/PlanService.java
@@ -1,5 +1,7 @@
 package team.seventhmile.tripforp.domain.plan.service;
 
+import java.util.List;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -11,27 +13,19 @@ import team.seventhmile.tripforp.domain.plan.dto.CreatePlanRequest;
 import team.seventhmile.tripforp.domain.plan.dto.CreatePlanResponse;
 import team.seventhmile.tripforp.domain.plan.dto.GetPlanListResponse;
 import team.seventhmile.tripforp.domain.plan.dto.GetPlanResponse;
-import team.seventhmile.tripforp.domain.plan.dto.PlanGetDetailDto;
 import team.seventhmile.tripforp.domain.plan.dto.PlanGetDto;
-import team.seventhmile.tripforp.domain.plan.dto.PlanGetItemDto;
-import team.seventhmile.tripforp.domain.plan.dto.PlanItemDto;
-import team.seventhmile.tripforp.domain.plan.dto.PlanLikeDto;
 import team.seventhmile.tripforp.domain.plan.dto.PlanListItemDto;
 import team.seventhmile.tripforp.domain.plan.dto.UpdatePlanItemRequest;
 import team.seventhmile.tripforp.domain.plan.dto.UpdatePlanRequest;
 import team.seventhmile.tripforp.domain.plan.dto.UpdatePlanResponse;
 import team.seventhmile.tripforp.domain.plan.entity.Area;
 import team.seventhmile.tripforp.domain.plan.entity.Plan;
-import team.seventhmile.tripforp.domain.plan.dto.UserGetDto;
-import team.seventhmile.tripforp.domain.plan.dto.AreaDto;
-import java.util.List;
-import java.util.stream.Collectors;
-
 import team.seventhmile.tripforp.domain.plan.repository.PlanLikeRepository;
 import team.seventhmile.tripforp.domain.plan.repository.PlanRepository;
 import team.seventhmile.tripforp.domain.user.entity.User;
 import team.seventhmile.tripforp.domain.user.repository.UserRepository;
 import team.seventhmile.tripforp.global.exception.ResourceNotFoundException;
+import team.seventhmile.tripforp.global.exception.UnauthorizedAccessException;
 
 @Service
 @Transactional(readOnly = true)
@@ -57,7 +51,6 @@ public class PlanService {
             .area(request.getArea())
             .build();
         planRepository.save(plan);
-        planRepository.flush();
 
         for (CreatePlanItemRequest planItemRequest : request.getPlanItems()) {
             //장소 가져오기, 없을 경우 장소 등록
@@ -67,9 +60,12 @@ public class PlanService {
     }
 
     @Transactional
-    public UpdatePlanResponse updatePlan(Long id, UpdatePlanRequest request) {
+    public UpdatePlanResponse updatePlan(Long id, UpdatePlanRequest request, UserDetails user) {
+
         Plan plan = planRepository.findById(id)
             .orElseThrow(() -> new ResourceNotFoundException(Plan.class, id));
+
+        checkPlanOwner(user, plan);
 
         plan.updatePlan(request);
 
@@ -81,9 +77,11 @@ public class PlanService {
     }
 
     @Transactional
-    public void deletePlan(Long id) {
+    public void deletePlan(Long id, UserDetails user) {
         Plan plan = planRepository.findById(id)
             .orElseThrow(() -> new ResourceNotFoundException(Plan.class, id));
+
+        checkPlanOwner(user, plan);
 
         planRepository.delete(plan);
     }
@@ -122,5 +120,14 @@ public class PlanService {
         plan.increaseViews();
 
         return new GetPlanResponse(plan, likeCount);
+    }
+
+    /**
+     * 수정, 삭제 시 작성자 본인이 맞는지 검증
+     */
+    private void checkPlanOwner(UserDetails user, Plan plan) {
+        if (!user.getUsername().equals(plan.getUser().getEmail())) {
+            throw new UnauthorizedAccessException(Plan.class);
+        }
     }
 }
